Move MirrorFilterUtils
